% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antMapR.R
\name{antMapR}
\alias{antMapR}
\title{antMapR function}
\usage{
antMapR(M, nCore = 1, ...)
}
\arguments{
\item{M}{[value]}

\item{nCore}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
