% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterM.R
\name{clusterM}
\alias{clusterM}
\title{clusterM function}
\usage{
clusterM(M, nExpChr, method = "hclust", nK = 1, makePlots = TRUE)
}
\arguments{
\item{M}{[value]}

\item{nExpChr}{[value]}

\item{method}{[value]. Default is "hclust"}

\item{nK}{[value]. Default is 1}

\item{makePlots}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
