% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{cross function}
\usage{
cross(geno1, geno2 = NULL, DH = FALSE, BC = FALSE)
}
\arguments{
\item{geno1}{[value]}

\item{geno2}{[value]. Default is NULL}

\item{DH}{[value]. Default is FALSE}

\item{BC}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
