% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{genotype}
\alias{genotype}
\title{genotype function}
\usage{
genotype(pop, loci = NULL, simplify = TRUE)
}
\arguments{
\item{pop}{[value]}

\item{loci}{[value]}

\item{simplify}{[value]}
}
\value{
[value]
}
\description{
function to genotype a population.
}
\details{
[fill in details here]
}
\examples{
none
}
