% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOlist.R
\name{getCOlist}
\alias{getCOlist}
\title{getCOlist function}
\usage{
getCOlist(Xl)
}
\arguments{
\item{Xl}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
