% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrderedM.R
\name{getOrderedM}
\alias{getOrderedM}
\title{getOrderedM function}
\usage{
getOrderedM(sol, Mgrp)
}
\arguments{
\item{sol}{[value]}

\item{Mgrp}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
