% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeq.R
\name{getSeq}
\alias{getSeq}
\title{getSeq function}
\usage{
getSeq(pop)
}
\arguments{
\item{pop}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
