% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqMatrix.R
\name{getSeqMatrix}
\alias{getSeqMatrix}
\title{getSeqMatrix function}
\usage{
getSeqMatrix(alCounts, chr = NULL)
}
\arguments{
\item{alCounts}{[value]}

\item{chr}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
