% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haldane.R
\name{haldane}
\alias{haldane}
\title{haldane function}
\usage{
haldane(r)
}
\arguments{
\item{r}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
