% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeF1.R
\name{makeF1}
\alias{makeF1}
\title{makeF1 function}
\usage{
makeF1(markerPos, nChrom = NULL, cM = NULL)
}
\arguments{
\item{markerPos}{[value]}

\item{nChrom}{[value]. Default is NULL}

\item{cM}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
