% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePop.R
\name{makePop}
\alias{makePop}
\title{makePop function}
\usage{
makePop(f1, popSize = 100, type = "BC", gen = 6)
}
\arguments{
\item{f1}{[value]}

\item{popSize}{[value]. Default is 100}

\item{type}{[value]. Default is "BC"}

\item{gen}{[value]. Default is 6}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
