% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mate.R
\name{mate}
\alias{mate}
\title{randomIntermate function}
\usage{
mate(genolist, newpopsize = length(genolist))
}
\arguments{
\item{genolist}{[value]}

\item{newpopsize}{[value]. Default is length(genolist)}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
