% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meiosis.R
\name{meiosis}
\alias{meiosis}
\title{meiosis function}
\usage{
meiosis(geno)
}
\arguments{
\item{geno}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
