% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderTocM.R
\name{orderTocM}
\alias{orderTocM}
\title{orderTocM function}
\usage{
orderTocM(sol, Mgrp)
}
\arguments{
\item{sol}{[value]}

\item{Mgrp}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
