% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLD.R
\name{plotLD}
\alias{plotLD}
\title{plotLD function}
\usage{
plotLD(M)
}
\arguments{
\item{M}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
