% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollinate.R
\name{pollinate}
\alias{pollinate}
\title{pollinate function}
\usage{
pollinate(gamete1, gamete2 = NULL)
}
\arguments{
\item{gamete1}{[value]}

\item{gamete2}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
