% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombine.R
\name{recombine}
\alias{recombine}
\title{recombine function}
\usage{
recombine(chrom, trackRecomb = FALSE, force1recomb = FALSE)
}
\arguments{
\item{chrom}{[value]}

\item{trackRecomb}{[value]. Default is FALSE}

\item{force1recomb}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
