% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recSel.R
\name{recSel}
\alias{recSel}
\title{recSel function}
\usage{
recSel(R, N, method = "uniRec", tR = TRUE)
}
\arguments{
\item{R}{[value]}

\item{N}{[value]}

\item{method}{[value]. Default is "uniRec"}

\item{tR}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
