% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMarkerPos.R
\name{sampleMarkerPos}
\alias{sampleMarkerPos}
\title{sampleMarkerPos function}
\usage{
sampleMarkerPos(nMarkers, cM = 100, nChrom = NULL,
  method = "uniform")
}
\arguments{
\item{nMarkers}{[value]}

\item{cM}{[value]. Default is 100}

\item{nChrom}{[value]. Default is NULL}

\item{method}{[value]. Default is "uniform"}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
