% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqInd.R
\name{seqInd}
\alias{seqInd}
\title{seqInd function}
\usage{
seqInd(ind)
}
\arguments{
\item{ind}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
