% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{cross function}
\usage{
cross(ind1, ind2 = NULL, DH = FALSE, BC = FALSE)
}
\arguments{
\item{ind1}{list. list of a single individual's chromosomes. Chromosomes should be matrices of logical values with 2 coulmns}

\item{ind2}{list. Same as ind1, if ind2 = NULL, then ind1 is crossed to itself (i.e. selfed).}

\item{DH}{logical. Should a doubled haploid be produced?}

\item{BC}{logical. Should a doubled haploid be produced?}
}
\value{
returns an individual resulting from a cross of ind1 and ind2.
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
