% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollinate.R
\name{pollinate}
\alias{pollinate}
\title{pollinate function}
\usage{
pollinate(gamete1, gamete2 = NULL)
}
\arguments{
\item{gamete1}{list. List of haploid chromosomes, where chromosomes have 1 column.}

\item{gamete2}{list. Same as gamete1. If gamete2 = NULL, then}
}
\value{
a list of chromosomes for 1 individual formed from the gametes provided.
}
\description{
Make individual from gametes.
}
\details{
none
}
\examples{
none
}
